backend/pyproject.toml
[project]
name = "family-app-backend"
version = "0.1.0"
description = "Family app backend"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.136.1",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.46.0",
]

[dependency-groups]
dev = [
    "async-asgi-testclient>=1.4.11",
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.12",
]

[tool.ruff]
target-version = "py314"
line-length = 120
exclude = [
    ".bzr", ".direnv", ".eggs", ".git", ".git-rewrite", ".hg",
    ".ipynb_checkpoints", ".mypy_cache", ".nox", ".pants.d", ".pyenv",
    ".pytest_cache", ".pytype", ".ruff_cache", ".svn", ".tox",
    ".venv", ".vscode", "__pypackages__", "_build", "buck-out",
    "build", "dist", "node_modules", "site-packages", "venv",
]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "B",    # flake8-bugbear
    "S",    # flake8-bandit
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "SIM",  # flake8-simplify
    "PTH",  # flake8-use-pathlib
    "RUF",  # Ruff-specific rules
]
ignore = [
    "E501",  # line-too-long (formatter handles this)
    "E203",
    "B008",  # FastAPI Depends() calls in function defaults
    "S101",  # assert statements (used in tests)
]
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
line-ending = "lf"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
